sunxi-mmc: Add new sunxi-mmc driver from a20 code dump
The sunxi-3.3-cubieboard2 branch has a new mmc driver, which is much cleaner,
properly using the existing mmc framework for many things, and offers more
features then the original driver we carry in sunxi-3.4.

Note this commit just adds the driver as is. The next commits will actually
hook it up to the build system and port it to sunxi-3.4 in sofar that is
needed.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>